chore(viz-entity): Restore class hierarchy between Abstract and implementation
After creating the AbstractCamelVisualEntiy class, the connection
between the BaseVisualCamelEntity and its implementation has been lost
due to the Abstract class not implementing said interface.

This commit updates the Abstract class to implement the interface and
also consolidates the `toVizNode` method since it's the same between
both `CamelRouteVisualEntity` and `KameletVisualEntity`.
