+++ packages/ui/src/models/visualization/flows/kamelet-visual-entity.ts
@@ -1,16 +1,12 @@
 import { AbstractCamelVisualEntity } from './abstract-camel-visual-entity';
 /* eslint-disable no-case-declarations */
-import { ProcessorDefinition } from '@kaoto-next/camel-catalog/types';
 import { IKameletDefinition, IKameletMetadata, IKameletSpec } from '../..';
 import { getCamelRandomId } from '../../../camel-utils/camel-random-id';
-import { NodeIconResolver } from '../../../utils';
 import { EntityType } from '../../camel/entities';
-import { BaseVisualCamelEntity, IVisualizationNode } from '../base-visual-entity';
-import { CamelComponentSchemaService } from './support/camel-component-schema.service';
 
-export class KameletVisualEntity extends AbstractCamelVisualEntity implements BaseVisualCamelEntity {
+export class KameletVisualEntity extends AbstractCamelVisualEntity {
   id: string;
-  type = EntityType.Kamelet;
+  readonly type = EntityType.Kamelet;
   spec: IKameletSpec;
   metadata: IKameletMetadata;
 
@@ -22,25 +18,12 @@ export class KameletVisualEntity extends AbstractCamelVisualEntity implements Ba
   }
 
   /** Internal API methods */
-  getId(): string {
-    return this.id;
-  }
-
   setId(routeId: string): void {
     this.id = routeId;
     this.metadata.name = this.id;
   }
 
-  toVizNode(): IVisualizationNode {
-    const rootNode = this.getVizNodeFromProcessor('from', {
-      processorName: 'from' as keyof ProcessorDefinition,
-      componentName: CamelComponentSchemaService.getComponentNameFromUri(this.spec.template.from!.uri),
-    });
-    rootNode.data.entity = this;
-
-    if (!this.spec.template.from?.uri) {
-      rootNode.data.icon = NodeIconResolver.getPlaceholderIcon();
-    }
-    return rootNode;
+  protected getRootUri(): string | undefined {
+    return this.spec.template.from?.uri;
   }
 }
